requirements.txt
py-cord==2.6.1
python-dotenv==1.0.1
requests==2.32.3
isort==5.13.2